simplify logic for jwt refresh & logout/in if the refresh token has expired, when order-confirmation is requested

+++ src/store/utils/utils.js
@@ -1,11 +1,13 @@
 import api from "../../api/axiosDefaults";
 
 /** Send logout request */
-const sendLogoutRequest = async (id, toast, dispatch) => {
-  console.log("logout req");
+const sendLogoutRequest = async (id, toast, dispatch, path) => {
   await api.post(`/auth/signout/${id}`);
-  dispatch({ type: "LOGOUT_USER" });
-  localStorage.setItem("auth", null);
+  console.log("not deleting auth");
+  if (path !== "/order-confirm") {
+    dispatch({ type: "LOGOUT_USER" });
+    localStorage.setItem("auth", null);
+  }
   // localStorage.setItem("cartItems", []);
   // dispatch({ type: "CLEAR_CART" });
 };
@@ -14,14 +16,12 @@ export const sendRefreshJwt = async (toast, path, dispatch, getState) => {
   console.log("refreshing JWT");
   try {
     let { data } = await api.post(`/auth/refreshtoken`);
-    console.log(data);
     if (data.message === "JWT has been regenerated.") {
-      console.log(data);
       return true;
     } else if (data.message === "Refresh Token has expired.") {
       let userId = getState().auth.user.id;
       // logout the user
-      await sendLogoutRequest(userId, toast, dispatch);
+      await sendLogoutRequest(userId, toast, dispatch, path);
       // set login dialog & destination path after login
       await dispatch({
         type: "SET_MODAL",
